new int type; malloc passes with -m32

+++ src/log.cpp
@@ -2,8 +2,8 @@
 
 namespace ANAlloc {
 
-uint8_t Log2Ceil(uint64_t value) {
-  uint8_t floored = Log2Floor(value);
+int Log2Ceil(UInt value) {
+  int floored = Log2Floor(value);
   if ((1UL << floored) != value) {
     return floored + 1;
   } else {
